Piotrek branch

+++ src/main/java/com/footballleague/league_organizer/controllers/LeagueController.java
@@ -2,6 +2,7 @@ package com.footballleague.league_organizer.controllers;
 
 import com.footballleague.league_organizer.entities.League;
 
+import com.footballleague.league_organizer.entities.Team;
 import com.footballleague.league_organizer.services.LeagueService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -24,10 +25,13 @@ public class LeagueController {
         return leagueService.getAllLeagues();
     }
     @GetMapping("/{id}")
-    public ResponseEntity<Optional<League>> getLeagueById(@PathVariable int id){
+    public ResponseEntity<Optional<League>> getLeagueById(@PathVariable Long id){
         Optional<League> league = leagueService.getLeague(id);
         return ResponseEntity.ok(league);
     }
+
+
+
     @PostMapping
     public ResponseEntity<League> createLeague(@RequestBody League league) {
         League savedLeague = leagueService.createLeague(league);
