+++ src/main/java/com/footballleague/league_organizer/services/LeagueService.java
@@ -1,6 +1,7 @@
 package com.footballleague.league_organizer.services;
 
 import com.footballleague.league_organizer.entities.League;
+import com.footballleague.league_organizer.entities.Team;
 import com.footballleague.league_organizer.repositories.LeagueRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -19,7 +20,8 @@ public class LeagueService {
     public List<League> getAllLeagues() {
         return leagueRepository.findAll();
     }
-    public Optional<League> getLeague(int leagueId) {
+
+    public Optional<League> getLeague(Long leagueId) {
         return leagueRepository.findById(leagueId);
     }
     public League createLeague(League league) {
